Create a fileset list component; also added webpack folder alias for easier importing of files and avoiding relative paths from now on
