+++ assets/jest.config.js
@@ -3,6 +3,7 @@ module.exports = {
   moduleNameMapper: {
     "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$":
       "<rootDir>/__mocks__/fileMock.js",
+    "^@js(.*)$": "<rootDir>/js$1",
   },
   setupFiles: ["jest-localstorage-mock"],
   setupFilesAfterEnv: ["./jest.setup.js"],
