+++ assets/js/mock-data/filesets.js
@@ -1,0 +1,48 @@
+export const mockFileSets = [
+  {
+    id: "45226a50-87ca-443e-bc05-f47884e14505",
+    role: "AM",
+    accessionNumber: "Voyager:2569254_FILE_0",
+    metadata: {
+      description: "inu-dil-9d35d0ba-a84b-4e0a-99e6-9c6b548a46db.tif",
+      originalFilename: "inu-dil-9d35d0ba-a84b-4e0a-99e6-9c6b548a46db.jpg",
+      label: "inu-dil-9d35d0ba-a84b-4e0a-99e6-9c6b548a46db.jpg",
+      location:
+        "s3://dev-preservation/45/22/6a/50/6be181760c0adb1f3425a0ae3438f3633b0baa9a6f74afa973c94ae6de6f45cb",
+      sha256:
+        "6be181760c0adb1f3425a0ae3438f3633b0baa9a6f74afa973c94ae6de6f45cb",
+    },
+  },
+  {
+    __typename: "FileSet",
+    id: "109b9a5c-3c6f-4a98-b98b-12402b871dc7",
+    role: "AM",
+    accessionNumber: "Voyager:2572813_FILE_0",
+    metadata: {
+      __typename: "FileSetMetadata",
+      description: "inu-dil-41913a91-037f-494b-9113-06004a8a98fb.tif",
+      originalFilename: "inu-dil-41913a91-037f-494b-9113-06004a8a98fb.jpg",
+      label: "inu-dil-41913a91-037f-494b-9113-06004a8a98fb.jpg",
+      location:
+        "s3://dev-preservation/10/9b/9a/5c/1477fbefbeeb04f0d02ac3cbd9594df0d9e7edca993ec076272d7fea67ab26a8",
+      sha256:
+        "1477fbefbeeb04f0d02ac3cbd9594df0d9e7edca993ec076272d7fea67ab26a8",
+    },
+  },
+  {
+    __typename: "FileSet",
+    id: "d414d3d4-b72c-49cc-b7cc-faa9bc0f256e",
+    role: "AM",
+    accessionNumber: "Voyager:2553177_FILE_0",
+    metadata: {
+      __typename: "FileSetMetadata",
+      description: "inu-dil-96e6d167-5022-42e7-9de7-7f851a866f44.tif",
+      originalFilename: "inu-dil-96e6d167-5022-42e7-9de7-7f851a866f44.jpg",
+      label: "inu-dil-96e6d167-5022-42e7-9de7-7f851a866f44.jpg",
+      location:
+        "s3://dev-preservation/d4/14/d3/d4/f7f1324a418e2c7c6ef17c45fc14ec6ce5a6124e636cb96272a7f35dc72d9664",
+      sha256:
+        "f7f1324a418e2c7c6ef17c45fc14ec6ce5a6124e636cb96272a7f35dc72d9664",
+    },
+  },
+];
