+++ assets/package.json
@@ -41,6 +41,7 @@
     "prop-types": "^15.7.2",
     "rc-progress": "^3.1.0",
     "react-apollo": "^3.1.5",
+    "react-beautiful-dnd": "^13.0.0",
     "react-csv": "^2.0.3",
     "react-hook-form": "^6.8.6",
     "react-json-pretty": "^2.2.0"
